Overcome "Connection Refused" on some operations
So some operations on Windows actually cause the WinRM service to become
unavailable for some time and WinRM currently cannot overcome this
problem. Since this is a real problem to us when e.g. installing SCVMM
on servers and we cannot influence the installer, here is the
work-around we implemented.

The current implementation simply tries 5 times with a 5 seconds delay,
but YMMV. Ideally this is configurable by the user, and future defaults
should be tested.